Prepping for push to GitHub.
The previous approach of only showing some contact information when
accessed with a "secret" path doesn't really work when the full source
is public. Used `git filter-branch` to wipe out the contact info from
history and moved it to a separate untracked file.

Now that I've done this it seems overly paranoid, but hey, I learned
how to use `git filter-branch`.